requirements.txt
python-telegram-bot==21.5
requests>=2.31.0
yt-dlp==2024.10.22